+++ src/contracts/interfaces/IFNFTCollectionFactory.sol
@@ -14,7 +14,7 @@ interface IFNFTCollectionFactory is IBeacon {
   function factoryTargetRedeemFee() external view returns (uint64);
   function factoryRandomSwapFee() external view returns (uint64);
   function factoryTargetSwapFee() external view returns (uint64);
-  function swapFee() external view returns (uint64);
+  function swapFee() external view returns (uint256);
   function vaultFees(uint256 vaultId) external view returns (uint256, uint256, uint256, uint256, uint256);
   function flashLoanFee() external view returns (uint64);
